requirements.txt
requests>=2.31
beautifulsoup4>=4.12
playwright>=1.40
openpyxl>=3.1
anthropic>=0.40
pyyaml>=6.0
rich>=13.0
